+++ src/rowData.ts
@@ -1,7 +1,7 @@
-import { Node } from "./node";
+import { InspectorNode } from "./inspectorNode";
 
 export class RowData {
-    node!: Node;
+    node!: InspectorNode;
     index!: number;
     lastUsed: number = Date.now();
 }